`@remotion/browser-studio`: Split a sequence at the playhead
