+++ packages/studio/src/components/split-jsx-sequence-api.ts
@@ -1,0 +1,15 @@
+import type {
+	SplitJsxSequenceRequest,
+	SplitJsxSequenceResponse,
+} from '@remotion/studio-shared';
+import {getBrowserStudioOperations} from '../helpers/browser-studio-operations';
+import {callApi} from './call-api';
+
+export const splitJsxSequence = (
+	request: SplitJsxSequenceRequest,
+): Promise<SplitJsxSequenceResponse> => {
+	const browserStudioOperations = getBrowserStudioOperations();
+	return browserStudioOperations
+		? browserStudioOperations.splitJsxSequence(request)
+		: callApi('/api/split-jsx-sequence', request);
+};
